Refactored; needs testing
